initial refactor, moved logs logic to sub-module logs
